requirements.txt
flask==2.3.2
flask-cors==3.0.10
numpy==1.24.3
opencv-python-headless==4.7.0.72
joblib==1.2.0
pillow==9.5.0
PyWavelets==1.4.1
scikit-learn==1.3.0
gunicorn==20.1.0